UBER-845: Change strip html tags implementation from SAX to Prosemirror model
